Now able to upload advanced tasks through AddEditToDoActivity Implemented viewholders for tasks with set date or repeating occurences.

+++ app/src/main/java/com/example/reminderapp/adapters/DoneTasksAdapter.kt
@@ -40,9 +40,11 @@ class DoneTasksAdapter(options : FirestoreRecyclerOptions<TaskModel>) : Firestor
         }
 
         holder.taskDelete.setOnClickListener{
-            db = FirebaseFirestore.getInstance()
-            db.collection(Constants.TasksCollection).document(model.taskID!!).delete()
-            notifyItemRemoved(position)
+            if(itemCount != 0){
+                db = FirebaseFirestore.getInstance()
+                db.collection(Constants.TasksCollection).document(model.taskID!!).delete()
+                notifyItemRemoved(position)
+            }
         }
     }
 }
